+++ XamlPowerToys/Infrastructure/AssemblyAssistant.cs
@@ -1,27 +1,22 @@
 ﻿namespace XamlPowerToys.Infrastructure {
+
     using System;
     using System.Globalization;
-
-    //using System.Globalization;
     using System.IO;
     using System.Linq;
-    //using System.Runtime.InteropServices;
     using EnvDTE;
     using Microsoft.Build.Construction;
-    //using Microsoft.VisualStudio.Shell.Interop;
     using XamlPowerToys.Model;
     using XamlPowerToys.UI.Infrastructure;
-    //using Constants = Microsoft.VisualStudio.OLE.Interop.Constants;
-    //using IServiceProvider = Microsoft.VisualStudio.OLE.Interop.IServiceProvider;
 
     internal class AssemblyAssistant {
-
         public const String XamarinProjectGuidString = "786C830F-07A1-408B-BD7F-6EE04809D6DB";
         public const String WpfProjectGuidString = "60DC8134-EBA5-43B8-BCC9-BB4BC16C2548";
         public const String UwpProjectGuidString = "A5A43C5B-DE2A-4C0C-9213-0A381AF9435A";
         public const String SilverlightProjectGuidString = "A1591282-1198-4647-A2B1-27E5FF5F6F3B";
 
         internal static String GetAssemblyPath(Project vsProject) {
+            Microsoft.VisualStudio.Shell.ThreadHelper.ThrowIfNotOnUIThread();
             var fullPath = Path.GetDirectoryName(vsProject.FullName);
             var outputPath = vsProject.ConfigurationManager.ActiveConfiguration.Properties.Item("OutputPath").Value.ToString();
 
@@ -50,6 +45,7 @@
         }
 
         internal static ProjectType GetProjectType(Project vsProject) {
+            Microsoft.VisualStudio.Shell.ThreadHelper.ThrowIfNotOnUIThread();
             var result = GetProjectTypeGuids(vsProject).ToUpper();
 
             if (result.Contains(AssemblyAssistant.WpfProjectGuidString)) {
@@ -69,6 +65,7 @@
         }
 
         internal static String GetProjectFrameworkVersion(Project vsProject) {
+            Microsoft.VisualStudio.Shell.ThreadHelper.ThrowIfNotOnUIThread();
             try {
                 return vsProject.Properties.Item("TargetFrameworkMoniker").Value.ToString().Split(',')[1].Replace("Version=v", String.Empty);
             } catch (Exception ex) {
@@ -78,6 +75,7 @@
         }
 
         internal static String GetProjectTypeGuids(Project vsProject) {
+            Microsoft.VisualStudio.Shell.ThreadHelper.ThrowIfNotOnUIThread();
             var projectRoot = ProjectRootElement.Open(vsProject.FileName);
             // ReSharper disable once PossibleNullReferenceException
             var netStandardTest = projectRoot.AllChildren.OfType<ProjectPropertyElement>().FirstOrDefault(x => x.Name == "TargetFramework" && x.Value.StartsWith("netstandard"));
@@ -98,19 +96,6 @@
             }
 
             return String.Empty;
-
-            //var ivsSolution = (IVsSolution)GetService(proj.DTE, typeof(IVsSolution));
-            //IVsHierarchy ivsHierarchy;
-            //Int32 result = ivsSolution.GetProjectOfUniqueName(proj.UniqueName, out ivsHierarchy);
-
-            //var projectTypeGuids = String.Empty;
-
-            //if (result == 0) {
-            //    var ivsAggregatableProject = (IVsAggregatableProject)ivsHierarchy;
-            //    ivsAggregatableProject.GetAggregateProjectTypeGuids(out projectTypeGuids);
-            //}
-
-            //return projectTypeGuids;
         }
 
         internal static Boolean SkipLoadingAssembly(String assemblyName) {
@@ -169,36 +154,23 @@
             if (assemblyName.StartsWith("prism")) {
                 return true;
             }
-            if (assemblyName.StartsWith("dryloc")) {
+            if (assemblyName.StartsWith("dryioc")) {
+                return true;
+            }
+            if (assemblyName.StartsWith("unity")) {
                 return true;
             }
             if (assemblyName.StartsWith("silverlight")) {
                 return true;
             }
+            if (assemblyName.StartsWith("fontawesome")) {
+                return true;
+            }
+            if (assemblyName.StartsWith("commonservicelocator")) {
+                return true;
+            }
+
             return false;
         }
-
-        //static Object GetService(Object serviceProvider, Guid guid) {
-        //    Object service = null;
-        //    IntPtr intPtr;
-        //    var guidService = guid;
-        //    var riidGuid = guidService;
-        //    var iServiceProvider = (IServiceProvider)serviceProvider;
-        //    var hr = iServiceProvider.QueryService(ref guidService, ref riidGuid, out intPtr);
-
-        //    if (hr != 0) {
-        //        Marshal.ThrowExceptionForHR(hr);
-        //    } else if (!intPtr.Equals(IntPtr.Zero)) {
-        //        service = Marshal.GetObjectForIUnknown(intPtr);
-        //        Marshal.Release(intPtr);
-        //    }
-
-        //    return service;
-        //}
-
-        //static Object GetService(Object serviceProvider, Type type) {
-        //    return GetService(serviceProvider, type.GUID);
-        //}
-
     }
 }