+++ XamlPowerToys/Properties/AssemblyInfo.cs
@@ -1,14 +1,15 @@
 ﻿using System.Reflection;
 using System.Runtime.InteropServices;
+using XamlPowerToys;
 
-[assembly: AssemblyTitle("XAML Power Toys for Visual Studio 2015 - 2017")]
-[assembly: AssemblyDescription("Visual Studio 2015 - 2017 extension that enables developers to rapidly generate data entry forms from view models or entity objects for Xamarin Forms, UWP, WPF, and Silverlight.")]
+[assembly: AssemblyTitle(Vsix.Name)]
+[assembly: AssemblyDescription(Vsix.Description)]
 [assembly: AssemblyConfiguration("")]
-[assembly: AssemblyCompany("Oceanware")]
-[assembly: AssemblyProduct("XAML Power Toys")]
-[assembly: AssemblyCopyright("Copyright © 2008-2017")]
+[assembly: AssemblyCompany(Vsix.Author)]
+[assembly: AssemblyProduct(Vsix.Name)]
+[assembly: AssemblyCopyright("Copyright © 2008-2019")]
 [assembly: AssemblyTrademark("")]
 [assembly: AssemblyCulture("")]
 [assembly: ComVisible(false)]
-[assembly: AssemblyVersion("1.5.2.*")]
-[assembly: AssemblyFileVersion("1.5.2.0")]
+[assembly: AssemblyVersion(Vsix.Version)]
+[assembly: AssemblyFileVersion(Vsix.Version)]