+++ CHANGELOG.md
@@ -10,6 +10,12 @@ download in the
 
 These are the changes to each version that has been released on the official Visual Studio extension gallery.
 
+## 2.0.0
+
+- [x] All Updated for Visual Studio 2017 and 2019
+- [x] Dropped support for Visual Studio 2015, limitation of VS Extensibility changes that support Async loading of extensions.
+
+
 ## 1.5.32
 
 - [x] All .ToLower() calls converted to .ToLower(CultureInfo.InvariantCulture).